+++ config/src/font.rs
@@ -271,9 +271,15 @@ bitflags! {
     }
 }
 
+impl FreeTypeLoadFlags {
+    pub fn default_hidpi() -> Self {
+        Self::NO_HINTING
+    }
+}
+
 impl Default for FreeTypeLoadFlags {
     fn default() -> Self {
-        Self::NO_HINTING
+        Self::DEFAULT
     }
 }