+++ docs/changelog.md
@@ -22,7 +22,9 @@ usually the best available version.
 As features stabilize some brief notes about them will accumulate here.
 
 #### Changed
-* Not yet!
+* The default for
+  [freetype_load_flags](config/lua/config/freetype_load_flags.md) is now
+  `NO_HINTING` when the dpi is >= 100, otherwise `DEFAULT`. #4902
 #### New
 #### Fixed
 * It was not possible to specify `freetype_load_flags = 'DEFAULT'`. #4902